tests, ping: Move ping under libnet package
