+++ tests/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
         "config.go",
         "job.go",
         "login.go",
-        "ping.go",
         "pod_servers.go",
         "test.go",
         "utils.go",
