+++ tests/libnet/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "cluster.go",
         "dns.go",
         "ipaddress.go",
+        "ping.go",
         "skips.go",
         "validation.go",
     ],
@@ -15,6 +16,7 @@ go_library(
     deps = [
         "//staging/src/kubevirt.io/client-go/api/v1:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
+        "//tests/console:go_default_library",
         "//tests/flags:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
         "//vendor/github.com/onsi/gomega:go_default_library",
